Implemented a simple memory overhead computation pass
Currently the memory overhead computation is an estimate: the mayLoad()
and mayStore() may return true for some non-load/store instructions.
For an accurate implementation, we need examine every instruction
individually.

We will do that once we're sure that the the shadow stack strategy works
and the memory overhead is acceptable.