Conversion of tSTRspi and tSTRi passed current tests
Added more test code. The whiltlist of functions are these we want to
run passes on, and the blacklist is for those functions on which we
don't want to run passes.

Currently instrumented SystemInit() function causes the program go into
an infinite loop. The stores converted in this function are common
stores whose unprivileged versions work fine in other functions. So,
my guess is that some stores in this function are supposed to be running
as privileged store. We need to verify this later.