Rewrote the pass to compute memory overhead
The old pass was very corse-grained: it assumed all the loads and stores
are 2 bytes long, while some of them are actually 4 bytes. It also
didn't count those add and sub instructions added to the program when
constructing unprivileged loads and stores.

Now the pass deals with each kind of load and store individually.