Added more funcs to the blacklist & moved func lists to a separate file.
We use timer related functions to record the execution time of benchmark
programs. Instrumenting those timer related functions breaks a test
program: the HAL_Init() function (the function we added to prepare the
timer) won't return. I guess that it's because these functions need to
run in privileged mode. We need to review this issue later.

Also moved the whitelist and blacklist of functions into a separate file
because the lists grew pretty quickly and putting them apart from the
header file of the pass makes code cleaner.