Fixed a bug in converting normal stores to unprivileged ones.
- Unprivileged load/store instructions only support 8-bit immediate,
ranging from 0 to 255; but a normal store can have an immediate as large
as 1023.  For a load/store whose immediate is greater than 255, we need
to use extra add and sub instructions. See the code for more details.

- Moved Silhouette passes prior to ARMConstantIslandPass ; otherwise
sometimes we could get an

    `error in backend: out of range pc-relative fixup value`

when llvm generates native asm code after running the transformation
pass. I think it’s because our transformation increases the size of a
lot of load/store instructions and adds some add/sub instructions, some
branch instructions cannot reach their destination (for example, the
`beq` instruction can only jump to a pc-relative destination within
range -256 to 254 bytes). The ARMConstantIslandPass updates the jump
table so that branch instructions can jump further as necessary.