Implemented converting VSTR to unprivileged stores
Implemented converting floating-point stores (VSTR) to unprivileged
stores. Because unprivileged store word (STRT) only accepts a
core-register as the source register while VSTR requires a FP register
to be the source register, we need some extra instructions to finish the
conversion. See the comment and code of function convertVSTR() for more
details.