Fixed a subtle bug in converting floating-point stores.
Previously we just picked r0 and r1 to be the transitional registers to
store the floating-point number(s). However, there is a chance that the
base register is r0 or r1, in which case the destination address would
be destroied even if we saved r0 and r1 onto the stack first. Now we
pick register(s) apart from the base register.