Modified convertVSTR to be more general.
The old implementation of this function assumes that the argument MI is
a VSTR. However, when we instrument VSTM or VPUSH, the FP stores are
implicit. The new version can take a VSTM or VPUSH as the MI, and handle
any FP store as long as we tell it we're storing a single-precision or
double-precision register. See the changed code for more details.