Implemented converting PUSH to STRT.
Implemented converting a PUSH to one or multiple STRT and an additional
SUB (update the SP). A push is essentially a STM. Maybe we can merge
this function with convertSTM().

See the code for more details.

At this point, we've implemented instrumenting all stores except inline
assembly, which needs further examination.