Fixed a bug in converting VSTR to STRT.
1. When converting a VSTR to a STRT, we used a push to back up some
integer register(s); here is a pitfall: the push itself could be
privileged if the program is running as privileged. So now we use
update-SP + STRT to replace push.

2. There is something very weird. When we use BuildMI() to create a push
or pop, if we pass r0 and r1 to the register list, everything is fine:
the generated instruction would be push/pop {r0, r1}. However, if we
pass r7 and r8 (or larger number registers), the generated instruction
would be push/pop{r7}. Is this a bug in LLVM 4.0?