Updated the code of recording the code size of machine functions.
The LLVM built-in STATISTIC reports the code size of a whole compilation
unit. However, there are some functions that are compiled but not
included in the generated binary code. The new code records the name of
a function and its code size. I'll write script to filter out those that
are eliminated in the linking stage.