Fixed a bug in the store transformation pass.
To convert a regular store to an unprivileged one, sometimes we need add
extra instructions, such as an extra pair of add and sub instructions to
help compute the target memory address. Those added instructions may
change the processor status flags as a side effect, which may
accidentally change the sementics of the program.

Now we insert an IT instruction before each newly added instruction
that would change the status flags if they're outside of an IT block.
This guarantees that the newly added 16-bit instructions will not update
the status flags. For those inserted 32-bit instruction that are able to
optionally update the status flags, currently these that are generated
by this pass don't have the "update flags" bit set on.

By the way, to fix this bug, I tried to "save status register (ARSR),
insert new instructions, and then restore the status register" but
failed. ARMv7-M doesn't allow us to push or store a special register to
memory. I also tried to "save a general-purpose register, move APSR to
the saved register, insert instructions, move from the saved register to
APSR, and then restore the saved register" but failed either. Although
ARMv7-M provides instructions (MSR and MRS) to move between regular
registers and special registers, I didn't manage to do that correctly.
For example, the following MI build

    BuildMI(......, TII->get(ARM::t2MSR_M), ARM::ARSR)
        .addReg(ARM::ARSR)
        .addReg(ARM::R3);

would generate assembly code

    msr IAPSR, r1

I haven't figured out the reason yet.