Fixed a bug in convertVSTR()
When picking a interim register, the old way is to find the next
register after the base register (e.g., if the base register is R0, then
we pick R1 as the interim register). However, as mentioned in
commit b291c90, LLVM 4.0 cannot push or
pop a register whose number is greater than R7. If the base reigster is
R7 and we pick R8 as the interim register, then the pop instruction
would be pop{}. So we just pick R0 or R1 as the interim register.