Added a function to split IT blocks with stores.
There is a bug to fix.

There are some store instructions that are within an IT block. The
transformation of these instructions may add extra instructions, such as
a pair of add and sub intructions; those added instructions should all
be conditionally executed as directed by the IT instruction. Besides,
the added instructions could affect other condtion instructions in the
IT block if there are at least two instructions in it: other intructions
may get squeezed outside the IT block by the newly added instructions.

This splitITBlockWithSTR() function splits an IT block into several ones
if it has at least one store instruction and has at least two
instructions in it.  It inserts an IT instruction for each of the
instruction inside the IT block and removes the original IT instruction.

Sepaprating all the instructions and their corresponding IT instruction
makes the implementation of the store transformation clearer.

The bug has not been completely fixed. We need use this function for
each store that is in an IT block.