Rewrote the code of handling imm that is greater than 255.
Now we put the piece of code that handles unprivileged stores whose imm
is greater than 255 in the convertSTRimm() function. It purified the
logic of buildUnprivStr().

The reason of this change is to simplify gathering all the added
instructions for a store transformation; otherwise the logic of
collecting all the inserted instructions of the convertSTRimm() would be
a little more complicated than other store conversion functions. And the
reason why we want to collect all the inserted instructions is that for
a store that is inside an IT block, we need to insert an IT instruction
before each added instruction that comes from the store transformation.
Gathering them together simplifies the implementation.