Added a function to add a tADDi8 or tSUBi8 instruction.
This function builds an tADDi8 or tSUBi8 instruction as part of building
an unprivileged store. The reason we need a whole function to build such
an
add/sub is that tADDi8 and tSUBi8 only supports R0 - R7 as the
source/destination register. Sometimes the base regisger of a store is
from R8 - R12; so we need to handle this case specially.

See the code for more details.

TO-DO: replace all the tADDi8 and tSUBi8 builders with this function.