Fixed a bug in building a add/sub with SP as the base register.
To build an add/sub (imm), if the base register is SP then the imm must
be a multiple of 4 to do . Although according to the manual, the
Encoding T4 of ADD (SP + imm) allows imm to be any value in the range o
0 - 4095, I couldn't find such an ADD in the .td file. So we use extra
instructions to transform a store whose base register is SP and the imm
is greater than 255. Algorithm: pick a register, back it up on the
stack, use this register to compute the target address, do the store,
and then restore the register. This costs two more instructions then
using Encoding T4 of ADD (SP + imm).

See the code for more details.

Now I have tested all BEEBS pragrams that can compile and run with the
original clang. They all can compile and run with the store
transformation pass.