Fixed a bug in converting VSTR.
For a VSTR, when the immediate is negative, LLVM does not encode it as
a negative directly; instead it encodes it as (256 - imm). For example,
when the real immediate is -1, we would get 257 from the MI's encoding
for the immediate. Thus, we need get the real immediate before the
conversion work.

See the code for more details.