Fixed a subtle bug in Store Promotion.
There was a subtle bug: when the base register of a store instruction
is sp and the immediate is greater than 255, there is a hazard
adding a pair of add/sub instructions around the store because
if after the add there is a hardware interrupt kicking in, the
stack would be corrupted due to the context saving procedure.

Now we pick up a register, backup it, and use this register
to compute the target address.