Updated the SFI implementation for the new board.
Updated the SFI implementation for the new board setting.  Specifically:

Previously we use the B-L475E-IOT01A2 board which has 96 KB SRAM from
0x2000,0000 to 0x2001,7fff (RAM) and from 0x1000,0000 to 0x1000,7fff
(RAM2), so we put the shadow stack at the end of RAM from 0x2000,4000
and limit the SRAM space that a normal store can access to be within
0x1000,0000 to 0x1000,3fff or 0x2000,0000 to 0x2000,3fff by bit-masking
the pointer with 0xc000,c000.

Now with the new STM32F469I-DISCO board that comes with a 16 MB SDRAM
remapped at 0x0 to 0xff,ffff, we make the following adjustments.
1. Move heap, stack, and shadow stack to SDRAM and leave global on RAM.
2. Put the heap and stack from 0x0 to 0x7f,ffff and put shadow stack
   from 0x80,0000 to 0xff,ffff.
3. Bit-mask the pointer with 0xc080,0000.

As the new mask cannot be encoded by a single Thumb Modified Immediate
Constant, we have to use two bit-masking instructions now.