Added support for a guard region between stack and shadow stack.
Added support for a guard region between regular stack and shadow stack
which guards underflows of shadow stack from writing to regular stack.
This commit adds a command-line option to specify the guard region size
(default 1 MB) and adds code to reject a program if the stack frame size
of a function is larger than the guard region size.