Updated licensing and project information.
Updated the LICENSE.TXT file to clarify that it applies to the
Silhouette project.

Added text to original LLVM files we modified to note that we
modified them and that the changes are from the University of
Rochester.

Added text to files we added noting that the University of Rochester
owns the copyright.

Added text to files we added or modified to note that the license
file is in the top-level llvm directory and not on the LLVM web
page.